Fixed editing under GNU sed (see commit text)
After some Googling, it transpires that the original edit was written with
BSD sed in mind. Unfortunately, this meant that running this with GNU sed doesn't
work.
So, I fixed this by dropping the -i parameter and rewriting the command
(I mean copying and pasting from SE) in POSIX-compliant syntax.

Source and further reading: https://stackoverflow.com/a/44877280