Added argument pair support, complete with sanity checking
This commit adds support for specifying options in the format of parameter=value
on the command line.
Sanity checking was also implemented, and the program duly parses the arguments
for malformed input (parameter= / =value / =).
This is in addition to the skeleton of a function that would load a
"customization file" that the user could use to define their own delimiter
, default answers, and so forth.
One caveat of the current positional argument parser (looks for = wherever it
occurs) is that it prevents us from using = in other places (such as in a file
name). Despite this being an edge case that the doesn't break the current
implementation, I did add a commented-out section to make it look for a --
prefix (GNU argument specification, like `ls --verbose` for example), which
should solve this. I didn't complete that till the end because I simply got
too bored with positional arguments for now.