Add proxy variables to the ones in customization file
This adds the proxy variables to the ones loaded by the customization file.
The program will print the final values of the variables stored after
loading.
Caveat: how will we save the default value of the proxy variable when
generating a customization file, when there won't be a default value?