Hide token from console output, add comment to token variable
It might not be appropriate in some conditions to print the token to the
console. Therefore, we would just inform the user that a token had been
read from the file, without printing the actual token.

Added a small note in the code next to the token variable explaining where to
set that variable.