Don't treat "cust_file" as an admin's name
Previously, if "edit" was specified followed immediately by "cust_file", the
latter would get treated as an admin's name (instead of leaving the bot open as
per expected behavior).
It turns out that this is because "cust_file" is being treated as an admin's
name. This commit fixes this to ignore "cust_file".
This is a temporary fix, the whole getAdmins() function will be refactored at a
later time.