Allow specifying a token and a link to a customization file in env variables
This is intended for Heroku deployments. Heroku allows the user to specify
environment variables when deploying an app. Therefore, we take advantage of
this to allow the user to specify their token and a direct link to a JSON
customization file, as if they were specifying it on the command line.

Relevant status and error messages are also printed.